BIg update
App is functioning and will retreive the closest status from a year ago
today (whatever that may be). There are definitely still edgecases that
could cause bugs and I haven't done, full compatibility testing, but as
a prototype it is done.

Background still doesn't update to match user's profile from a year ago.